Add modal to add contact
Updated add contact button to bring down a modal. Adding contacts from the modal works the same as before. Added opacity to the background of modals for update and add contacts. @maxgomer , the stylesheet was referencing a class, but I caught it and changed the dots to hash tags to reference the ids. I used rgba to allow for opacity instead of the hexadecimal color. Slight modification the JS to allow the modal to disappear after adding the contact successfully.